Share common types and conversions with environment v1beta1
